+++ tests/test_main_command.py
@@ -8,8 +8,9 @@ if TYPE_CHECKING:
     from click.testing import CliRunner
     from pytest_mock import MockerFixture
 
-_LINUX_ONLY_COMMANDS = ('clean-old-kernels-modules', 'connect-g603', 'inhibit-notifications',
-                        'kill-gamescope', 'systemd-reset-tpm-cryptenroll', 'wait-for-disc')
+_LINUX_ONLY_COMMANDS = ('clean-old-kernels-modules', 'connect-g603', 'fix-mime-assocs',
+                        'inhibit-notifications', 'kill-gamescope', 'systemd-reset-tpm-cryptenroll',
+                        'wait-for-disc')
 _WINE_COMMANDS = ('kill-wine', 'set-wine-fonts', 'unregister-wine-assocs', 'winegoginstall',
                   'wineshell')